backend/requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart>=0.0.20

# Database & ORM
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache & Queue
redis==5.0.1

# Configuration & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# HTTP client
httpx==0.26.0

# LLM & RAG (chromadb/langchain은 DEMO_MODE=false + RAG 사용 시 별도 설치)
# chromadb>=0.4.22
# langchain>=0.1.0
# langchain-anthropic>=0.1.0
anthropic>=0.16.0

# Google Sheets integration
gspread==5.12.0
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Scheduler
apscheduler==3.10.4

# Authentication
PyJWT>=2.8.0
bcrypt>=4.0.0

# Rate Limiting
slowapi>=0.1.9
